[refactor] memberController에 dto 패키지 추가
- 기존 SignUpRequest는 controller 패키지에 들어있었으나, 이를 dto 패키지를 따로 생성하여 위치를 옮겨줌
